feat: integrate Gemini AI for figure generation and update API endpoints
